requirements.txt
python-docx>=1.1.0
PyYAML>=6.0
qdrant-client>=1.7.0
fastembed>=0.3.0
anthropic>=0.40.0
